# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "One Medy API Documentation",
        "description": "One Medy APi Service to be consumed by client Application",
        "contact": {
            "email": "info@walexbiznig.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/labtestrequest/add": {
            "post": {
                "tags": [
                    "Labtest Request"
                ],
                "summary": "Create a new test request to the Labs.",
                "operationId": "cfc80e17f5ffaaf3b9a6a7e0d1c78201",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateLabRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/v1/labtestrequests": {
            "get": {
                "tags": [
                    "Labtest Request"
                ],
                "summary": "Return a json array of all the tests requests sent to the labs.",
                "description": "Returns all test requests both the completed and uncompleted sent to the Laboratory. If iscompleted is placed in the URL as parameter with a value of\n     * 1, it will return all completed tests. The opposite happens if the value of iscompleted is made 0",
                "operationId": "getAllTestRequests",
                "parameters": [
                    {
                        "name": "iscompleted",
                        "in": "parameter",
                        "description": "Status of the Lab test",
                        "required": false
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/api/v1/pharmacy/requests/all": {
            "get": {
                "tags": [
                    "Pharmacy"
                ],
                "summary": "Get list of pharmacy requests",
                "description": "Returns list all pharmacy requests",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Size"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/doctor/rate/{id}": {
            "post": {
                "tags": [
                    "Rating"
                ],
                "summary": "rate a doctor",
                "operationId": "fd1503ce035d5941342df32b07b079c2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRatingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
        "/api/v1/tenant/create": {
            "post": {
                "tags": [
                    "Tenant"
                ],
                "summary": "register a Tenant",
                "operationId": "c6dd8b5fa113bb112d7e5954f6c976f4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTenantRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/v1/tenants": {
            "get": {
                "tags": [
                    "Tenant"
                ],
                "summary": "Get list of tenants",
                "description": "Returns list of tenants",
                "operationId": "getTenants",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Size"
                    },
                    {
                        "name": "tenanttypeid",
                        "in": "query",
                        "description": "the Id of the tenanttype"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/tenant/{id}": {
            "get": {
                "tags": [
                    "Tenant"
                ],
                "summary": "To get a particular tenant",
                "operationId": "bde219ee84137c680cde940fc86c4780",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tenant id",
                        "required": true
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/v1/tenant/{id}/status": {
            "patch": {
                "tags": [
                    "Administrator"
                ],
                "summary": "change the status of a tenant",
                "operationId": "594b4a50e2a85751d9af8a6d36e2af83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tenant Id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeTenantStatus"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateLabRequest": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "labtest_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "tenant_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "delivery_type": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "integer": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/CreateLabRequest"
                    }
                ]
            },
...
            "CreateRatingRequest": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "rating": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "user_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "rated_user_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/CreateRatingRequest"
                    }
                ]
            },
...
            "ChangeTenantStatus": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "status": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/ChangeTenantStatus"
                    }
                ]
            },
...
            "CreateTenantRequest": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "email": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "phonenumber": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "roles_id": {
                                "type": "integer",
                                "format": "int32"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "firstname": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "lastname": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "city": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "address": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "name": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "state_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "tenant_type_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/CreateTenantRequest"
                    }
                ]
            },
...
        },
...
    },
...
    "security": [
        []
    ]
}